park-here/backend/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Cors==4.0.0
Flask-JWT-Extended==4.5.2
python-dotenv==1.0.0
Flask-SocketIO==5.3.5
python-dateutil==2.8.2
PyJWT==2.8.0
requests==2.31.0
Werkzeug==2.3.7
